IGNITE-24846 Move typedef aliases to ignite-commons
