+++ modules/commons/src/main/java/org/apache/ignite/internal/util/typedef/C1.java
@@ -17,15 +17,15 @@
 
 package org.apache.ignite.internal.util.typedef;
 
-import org.apache.ignite.internal.util.lang.GridFunc;
-import org.apache.ignite.lang.IgniteInClosure;
+import org.apache.ignite.lang.IgniteClosure;
 
 /**
- * Defines {@code alias} for {@link org.apache.ignite.lang.IgniteInClosure} by extending it. Since Java doesn't provide type aliases
+ * Defines {@code alias} for {@link org.apache.ignite.lang.IgniteClosure} by extending it. Since Java doesn't provide type aliases
  * (like Scala, for example) we resort to these types of measures. This is intended to provide for more
- * concise code in cases when readability won't be sacrificed. For more information see {@link org.apache.ignite.lang.IgniteInClosure}.
- * @param <T> Type of the factory closure.
+ * concise code in cases when readability won't be sacrificed. For more information see {@link org.apache.ignite.lang.IgniteClosure}.
+ * @param <E1> Type of the free variable, i.e. the element the closure is called or closed on.
+ * @param <R> Type of the closure's return value.
  * @see GridFunc
- * @see org.apache.ignite.lang.IgniteInClosure
+ * @see org.apache.ignite.lang.IgniteClosure
  */
-public interface CI1<T> extends IgniteInClosure<T> { /* No-op. */ }
+public interface C1<E1, R> extends IgniteClosure<E1, R> { /* No-op. */ }
