+++ modules/commons/src/main/java/org/apache/ignite/internal/util/typedef/C2.java
@@ -17,16 +17,16 @@
 
 package org.apache.ignite.internal.util.typedef;
 
-import org.apache.ignite.internal.util.lang.GridFunc;
-import org.apache.ignite.lang.IgniteBiPredicate;
+import org.apache.ignite.lang.IgniteBiClosure;
 
 /**
- * Defines {@code alias} for {@link org.apache.ignite.lang.IgniteBiPredicate} by extending it. Since Java doesn't provide type aliases
+ * Defines {@code alias} for {@link org.apache.ignite.lang.IgniteBiClosure} by extending it. Since Java doesn't provide type aliases
  * (like Scala, for example) we resort to these types of measures. This is intended to provide for more
- * concise code in cases when readability won't be sacrificed. For more information see {@link org.apache.ignite.lang.IgniteBiPredicate}.
- * @param <T1> Type of the first free variable, i.e. the element the closure is called on.
- * @param <T2> Type of the second free variable, i.e. the element the closure is called on.
+ * concise code in cases when readability won't be sacrificed. For more information see {@link org.apache.ignite.lang.IgniteBiClosure}.
+ * @param <E1> Type of the first free variable, i.e. the element the closure is called or closed on.
+ * @param <E2> Type of the second free variable, i.e. the element the closure is called or closed on.
+ * @param <R> Type of the closure's return value.
  * @see GridFunc
- * @see org.apache.ignite.lang.IgniteBiPredicate
+ * @see org.apache.ignite.lang.IgniteBiClosure
  */
-public interface P2<T1, T2> extends IgniteBiPredicate<T1, T2> { /* No-op. */ }
+public interface C2<E1, E2, R> extends IgniteBiClosure<E1, E2, R> { /* No-op. */ }
