+++ modules/commons/src/main/java/org/apache/ignite/internal/util/typedef/CI1.java
@@ -17,15 +17,14 @@
 
 package org.apache.ignite.internal.util.typedef;
 
-import org.apache.ignite.internal.util.lang.GridFunc;
-import org.apache.ignite.lang.IgnitePredicate;
+import org.apache.ignite.lang.IgniteInClosure;
 
 /**
- * Defines {@code alias} for {@link org.apache.ignite.lang.IgnitePredicate} by extending it. Since Java doesn't provide type aliases
+ * Defines {@code alias} for {@link org.apache.ignite.lang.IgniteInClosure} by extending it. Since Java doesn't provide type aliases
  * (like Scala, for example) we resort to these types of measures. This is intended to provide for more
- * concise code in cases when readability won't be sacrificed. For more information see {@link org.apache.ignite.lang.IgnitePredicate}.
- * @param <E1> Type of the free variable, i.e. the element the predicate is called on.
+ * concise code in cases when readability won't be sacrificed. For more information see {@link org.apache.ignite.lang.IgniteInClosure}.
+ * @param <T> Type of the factory closure.
  * @see GridFunc
- * @see org.apache.ignite.lang.IgnitePredicate
+ * @see org.apache.ignite.lang.IgniteInClosure
  */
-public interface P1<E1> extends IgnitePredicate<E1> { /* No-op. */ }
+public interface CI1<T> extends IgniteInClosure<T> { /* No-op. */ }
