+++ modules/commons/src/main/java/org/apache/ignite/internal/util/typedef/CO.java
@@ -17,16 +17,14 @@
 
 package org.apache.ignite.internal.util.typedef;
 
-import org.apache.ignite.internal.util.lang.GridFunc;
-import org.apache.ignite.lang.IgniteReducer;
+import org.apache.ignite.lang.IgniteOutClosure;
 
 /**
- * Defines {@code alias} for {@link org.apache.ignite.lang.IgniteReducer} by extending it. Since Java doesn't provide type aliases
+ * Defines {@code alias} for {@link org.apache.ignite.lang.IgniteOutClosure} by extending it. Since Java doesn't provide type aliases
  * (like Scala, for example) we resort to these types of measures. This is intended to provide for more
- * concise code in cases when readability won't be sacrificed. For more information see {@link org.apache.ignite.lang.IgniteReducer}.
- * @param <E1> Type of the free variable, i.e. the element the closure is called or closed on.
- * @param <R> Type of the closure's return value.
+ * concise code in cases when readability won't be sacrificed. For more information see {@link org.apache.ignite.lang.IgniteOutClosure}.
+ * @param <T> Type of the factory closure.
  * @see GridFunc
- * @see org.apache.ignite.lang.IgniteReducer
+ * @see org.apache.ignite.lang.IgniteOutClosure
  */
-public interface R1<E1, R> extends IgniteReducer<E1, R> { /* No-op. */ }
+public interface CO<T> extends IgniteOutClosure<T> { /* No-op. */ }
