+++ modules/commons/src/main/java/org/apache/ignite/internal/util/typedef/P2.java
@@ -17,16 +17,15 @@
 
 package org.apache.ignite.internal.util.typedef;
 
-import org.apache.ignite.internal.util.lang.GridFunc;
-import org.apache.ignite.lang.IgniteBiInClosure;
+import org.apache.ignite.lang.IgniteBiPredicate;
 
 /**
- * Defines {@code alias} for {@link org.apache.ignite.lang.IgniteBiInClosure} by extending it. Since Java doesn't provide type aliases
+ * Defines {@code alias} for {@link org.apache.ignite.lang.IgniteBiPredicate} by extending it. Since Java doesn't provide type aliases
  * (like Scala, for example) we resort to these types of measures. This is intended to provide for more
- * concise code in cases when readability won't be sacrificed. For more information see {@link org.apache.ignite.lang.IgniteBiInClosure}.
- * @param <E1> Type of the first parameter.
- * @param <E2> Type of hte second parameter.
+ * concise code in cases when readability won't be sacrificed. For more information see {@link org.apache.ignite.lang.IgniteBiPredicate}.
+ * @param <T1> Type of the first free variable, i.e. the element the closure is called on.
+ * @param <T2> Type of the second free variable, i.e. the element the closure is called on.
  * @see GridFunc
- * @see org.apache.ignite.lang.IgniteBiInClosure
+ * @see org.apache.ignite.lang.IgniteBiPredicate
  */
-public interface CI2<E1, E2> extends IgniteBiInClosure<E1, E2> { /* No-op. */ }
+public interface P2<T1, T2> extends IgniteBiPredicate<T1, T2> { /* No-op. */ }
