+++ modules/commons/src/main/java/org/apache/ignite/internal/util/typedef/PKV.java
@@ -17,17 +17,14 @@
 
 package org.apache.ignite.internal.util.typedef;
 
-import org.apache.ignite.internal.util.lang.GridFunc;
-import org.apache.ignite.lang.IgniteBiClosure;
+import org.apache.ignite.lang.IgniteBiPredicate;
 
 /**
- * Defines {@code alias} for {@link org.apache.ignite.lang.IgniteBiClosure} by extending it. Since Java doesn't provide type aliases
- * (like Scala, for example) we resort to these types of measures. This is intended to provide for more
- * concise code in cases when readability won't be sacrificed. For more information see {@link org.apache.ignite.lang.IgniteBiClosure}.
- * @param <E1> Type of the first free variable, i.e. the element the closure is called or closed on.
- * @param <E2> Type of the second free variable, i.e. the element the closure is called or closed on.
- * @param <R> Type of the closure's return value.
+ * Defines {@code alias} for <tt>GridPredicate2&lt;K, V&gt;</tt> by extending
+ * {@link org.apache.ignite.lang.IgnitePredicate}. Since Java doesn't provide type aliases (like Scala, for example) we resort
+ * to these types of measures. This is intended to provide for more concise code without sacrificing
+ * readability. For more information see {@link org.apache.ignite.lang.IgnitePredicate}.
+ * @see org.apache.ignite.lang.IgniteBiPredicate
  * @see GridFunc
- * @see org.apache.ignite.lang.IgniteBiClosure
  */
-public interface C2<E1, E2, R> extends IgniteBiClosure<E1, E2, R> { /* No-op. */ }
+public interface PKV<K, V> extends IgniteBiPredicate<K, V> { /* No-op. */ }
